Refactored CI. (#829)
