+++ CMakeLists.txt
@@ -124,7 +124,7 @@ if(NOT OQS_BUILD_ONLY_LIB)
     if(NOT WIN32)
         add_custom_target(
             prettyprint
-            COMMAND find src tests .CMake -name '*.[ch]' | grep -v '/external/' | grep -v 'kem.*/pqclean_' | grep -v 'sig.*/pqclean_' | xargs astyle --options=.astylerc
+            COMMAND ./scripts/run_astyle.sh
             WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}
             USES_TERMINAL)
     endif()
