+++ tests/CMakeLists.txt
@@ -46,9 +46,6 @@ if(NOT WIN32)
     set(UNIX_TESTS test_aes test_hash test_sha3)
 
     set(PYTHON3_EXEC python3)
-    if(BUILD_SHARED_LIBS)
-        set(USE_LIBOQS_SO --use_liboqs_so)
-    endif()
 else()
     set(PYTHON3_EXEC python)
 endif()
@@ -89,14 +86,6 @@ endif()
 add_executable(speed_sig speed_sig.c)
 target_link_libraries(speed_sig PRIVATE ${API_TEST_DEPS})
 
-# See https://github.com/open-quantum-safe/liboqs/issues/738#issuecomment-621394744
-string(TOUPPER "$ENV{RUN_TESTS_PARALLEL}" RUN_TESTS_PARALLEL_UPPER)
-if(RUN_TESTS_PARALLEL_UPPER STREQUAL "FALSE")
-    set(NUMPROCESSES 1)
-else()
-    set(NUMPROCESSES auto)
-endif()
-
 # Record compile options -- from target speed_kem - don't set any options only for speed_kem!
 get_property(OQS_COMPILE_OPTIONS TARGET speed_kem PROPERTY COMPILE_OPTIONS)
 add_definitions(-DOQS_COMPILE_OPTIONS="[${OQS_COMPILE_OPTIONS}]")
@@ -106,7 +95,7 @@ add_definitions(-DOQS_COMPILE_OPTIONS="[${OQS_COMPILE_OPTIONS}]")
 # for DLL builds.
 add_custom_target(
     run_tests
-    COMMAND ${PYTHON3_EXEC} -m pytest --verbose --numprocesses=${NUMPROCESSES} --build_tool=${CMAKE_MAKE_PROGRAM} ${USE_LIBOQS_SO} --junitxml=${CMAKE_BINARY_DIR}/test-results/pytest/test-results.xml
+    COMMAND ${PYTHON3_EXEC} -m pytest --verbose --numprocesses=auto
     WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}
     DEPENDS oqs example_kem kat_kem test_kem example_sig kat_sig test_sig ${UNIX_TESTS}
     USES_TERMINAL)
